+++ src/irvine/oeis/a207/A207136.java
@@ -13,6 +13,6 @@ public class A207136 extends LambdaSequence {
 
   /** Construct the sequence. */
   public A207136() {
-    super(0, n -> Integers.SINGLETON.sum(0, n, k -> Binomial.binomial(n * n, k * (n - k))));
+    super(0, n -> Integers.SINGLETON.sum(0, n, k -> Binomial.binomial((long) n * n, (long) k * (n - k))));
   }
 }
